trf2csv working with new setup
