--- app/src/python/dicom_anonymise.py
@@ -1,10 +1,0 @@
-from pymedphys_dicom.dicom import DicomBase
-
-dicom = DicomBase.from_dict({
-    'Manufacturer': 'PyMedPhys',
-    'PatientName': 'Python^Monte'
-})
-
-dicom.anonymise(inplace=True)
-
-print(dicom.dataset)
