+++ app/src/python/init.tsx
@@ -1,10 +1,12 @@
 import { BehaviorSubject } from 'rxjs';
 
-import { pythonReady, IPythonData, pythonData } from '../observables/python';
+import {
+  pythonReady, IPythonData, pythonData, pythonCode
+} from '../observables/python';
 import { inputDirectory, outputDirectory } from '../observables/directories';
 
-import loadWheels from './load_wheels.py';
-import setupDirectories from './setup_directories.py';
+import loadWheels from './load-wheels.py';
+import setupDirectories from './setup-directories.py';
 
 declare let pyodide: any;
 declare let languagePluginLoader: any;
@@ -12,12 +14,14 @@ declare let languagePluginLoader: any;
 declare global {
   interface Window {
     pythonData: BehaviorSubject<IPythonData>;
+    pythonCode: BehaviorSubject<string>;
     inputDirectory: BehaviorSubject<Set<string>>;
     outputDirectory: BehaviorSubject<Set<string>>;
   }
 }
 
 window.pythonData = pythonData;
+window.pythonCode = pythonCode;
 window.inputDirectory = inputDirectory;
 window.outputDirectory = outputDirectory;
