+++ app/src/python/setup-directories.py
