+++ packages/pymedphys/src/pymedphys/trf.py
@@ -2,4 +2,4 @@ from pymedphys_fileformats.trf import (
     delivery_data_from_logfile, decode_header_from_file,
     trf2csv_by_directory)
 
-from pymedphys_logfiles.elekta import identify_logfile
+# from pymedphys_logfiles.elekta import identify_logfile
